Почти готово

+++ SecondProject_Auto/Contexts/AutoContext.cs
@@ -22,10 +22,10 @@ namespace SecondProject_Auto
         public QualityType QualityType { get; set; }
         public decimal Price { get; set; }
         public string Color { get; set; }
-        public uint Mileage { get; set; }
+        public float Mileage { get; set; }
         public FuelType FuelType { get; set; }
         public string Engine { get; set; }
-        public uint HorsePower { get; set; }
+        public int HorsePower { get; set; }
         public byte SeatingCapacity { get; set; }
         public string Mfr { get; set; }
         public string PhotoFilePath { get; set; }
